block all updates from Luma3DS 7.1
