+++ source/update.h
@@ -2,23 +2,15 @@
 
 #include "libs.h"
 
+#include "main.h"
 #include "release.h"
 
 #define MAXPATHLEN 37
-
-struct UpdateArgs {
-	PayloadType  payloadType;    /*!< Type of payload to upgrade  */
-	std::string  payloadPath;    /*!< Path to Luma3DS payload     */
-	bool         backupExisting; /*!< Backup existing payload     */
-	bool         migrateARN;     /*!< Migrate from AuReiNand      */
-	ReleaseVer   chosenVersion;  /*!< Version to update to        */
-	bool         isHourly;       /*!< Is chosen version a hourly? */
-};
 
 struct UpdateResult {
 	bool        success; /*!< Wether the operation was a success */
 	std::string errcode; /*!< Error code if success is false     */
 };
 
-UpdateResult update(const UpdateArgs& args);
-UpdateResult restore(const UpdateArgs& args);
+UpdateResult update(const UpdateInfo& args);
+UpdateResult restore(const UpdateInfo& args);
